urshin/pyproject.toml
[project]
name = "urshin"
version = "0.1.0"
description = "Generate urshi manifests from script URLs using LLM inference"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.9.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
urshin = "urshin.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["urshin"]
